add CarRepositoryTest

+++ src/main/resources/db/scripts/005_ddl_add_table_participates.sql
@@ -1,6 +1,6 @@
 CREATE TABLE participates (
                               id SERIAL PRIMARY KEY,
-                              post_id INT NOT NULL REFERENCES auto_post(id),
-                              user_id INT NOT NULL REFERENCES auto_user(id),
+                              post_id BIGINT NOT NULL REFERENCES auto_post(id),
+                              user_id BIGINT NOT NULL REFERENCES auto_user(id),
                               UNIQUE (post_id, user_id)
 );
